+++ src/main/java/mrriegel/storagenetwork/item/remote/ContainerRemote.java
@@ -44,11 +44,7 @@ public class ContainerRemote extends ContainerNetworkBase {
         this.addSlotToContainer(new Slot(matrix, index++, 8 + j * 18, 110 + i * 18));
       }
     }
-    for (int i = 0; i < 3; ++i) {
-      for (int j = 0; j < 9; ++j) {
-        this.addSlotToContainer(new Slot(playerInv, j + i * 9 + 9, 8 + j * 18, 174 + i * 18));
-      }
-    }
+    bindPlayerInvo(playerInv);
     for (int i = 0; i < 9; ++i) {
       if (i == playerInv.currentItem)
         this.addSlotToContainer(new Slot(playerInv, i, 8 + i * 18, 232) {
@@ -128,12 +124,6 @@ public class ContainerRemote extends ContainerNetworkBase {
     for (int i = 0; i < 9; i++) {
       NBTHelper.setItemStack(remoteItemStack, "c" + i, matrix.getStackInSlot(i));
     }
-  }
-
-  @Override
-  public void onContainerClosed(EntityPlayer playerIn) {
-    slotChanged();
-    super.onContainerClosed(playerIn);
   }
 
   @Override
